refactor(agents): migrate agent modules to TypeScript with shims
Address cv review feedback on PR #1618:
- Move agent-defs, agent-onboard, agent-runtime from bin/lib/ (CJS)
  to src/lib/ (TypeScript) following the established migration pattern
- Leave thin re-export shims in bin/lib/ pointing to dist/lib/
- Bump Hermes plugin version to 0.0.11
