+++ agents/hermes/plugin/plugin.yaml
@@ -2,7 +2,7 @@
 # SPDX-License-Identifier: Apache-2.0
 
 name: nemoclaw
-version: "0.1.0"
+version: "0.0.11"
 description: "NemoClaw sandbox integration for Hermes running inside OpenShell"
 author: "NVIDIA Corporation"
 manifest_version: 1
